changed float into uint32_t for visits counter fixes https://github.com/QueensGambit/CrazyAra/issues/39 reformulated connecting new node to search tree

+++ engine/src/evalinfo.cpp
@@ -109,6 +109,6 @@ void update_eval_info(EvalInfo& evalInfo, Node* rootNode, size_t tbHits, size_t 
         evalInfo.bestMoveQ = rootNode->get_q_value(bestMoveIdx);
         evalInfo.centipawns = value_to_centipawn(evalInfo.bestMoveQ);
     }
-    evalInfo.nodes = size_t(get_node_count(rootNode));
+    evalInfo.nodes = get_node_count(rootNode);
     evalInfo.tbHits = tbHits;
 }
